+++ st_link_analysis/frontend/src/components/infopanel.js
@@ -1,12 +1,12 @@
 import State from "../utils/state";
 
 // Constants / Configurations
-const INFOBAR_ID = "infobar";
-const LABEL_ID = "infobarLabel";
-const PROPS_ID = "infobarProps";
+const INFOPANEL_ID = "infopanel";
+const LABEL_ID = "infopanelLabel";
+const PROPS_ID = "infopanelProps";
 const NODEACTIONS_ID = "nodeActions";
 
-// Infobar children updates
+// Infopanel children updates
 function _updateLabel(color, label, icon) {
     const label_div = document.getElementById(LABEL_ID);
     label_div.firstChild.innerText = label;
@@ -27,17 +27,17 @@ function _updateProps(data) {
         })
         .map(([key, value]) => {
             return `
-        <div class='infobar__prop'>
-            <p class='infobar__key'>${key}</p>
-            <p class='infobar__val'>${value}</p>
+        <div class='infopanel__prop'>
+            <p class='infopanel__key'>${key}</p>
+            <p class='infopanel__val'>${value}</p>
         </div>`;
         })
         .join("");
 }
 
-// Infobar update
-function updateInfobar() {
-    const infobar = document.getElementById(INFOBAR_ID);
+// infopanel update
+function updateInfopanel() {
+    const infopanel = document.getElementById(INFOPANEL_ID);
     const nodeActions = document.getElementById(NODEACTIONS_ID);
     const el = State.getState("selection").lastSelected;
     let color, data, label, expanded, icon;
@@ -54,10 +54,10 @@ function updateInfobar() {
         expanded = false;
         icon = null;
     }
-    infobar.setAttribute("data-expanded", expanded);
+    infopanel.setAttribute("data-expanded", expanded);
     nodeActions.setAttribute("data-expanded", expanded);
     _updateLabel(color, label, icon);
     _updateProps(data);
 }
 
-export default updateInfobar;
+export default updateInfopanel;